src/org/launchcode/techjobs_oo/Job.java
package org.launchcode.techjobs_oo;

import java.util.Objects;

public class Job {

    private int id;
    private static int nextId = 1;

    private String name;
    private Employer employer;
    private Location location;
    private PositionType positionType;
    private CoreCompetency coreCompetency;


    //added constructors for unique ID and initializing fields

    public Job() {
        this.id = nextId;
        nextId++;
    }

    public Job(String name, Employer employer, Location location, PositionType positionType, CoreCompetency coreCompetency) {
        this();
        this.name = name;
        this.employer = employer;
        this.location = location;
        this.positionType = positionType;
        this.coreCompetency = coreCompetency;
    }

    //toString method for passing testForBlankLine
    public String toString() {
        return " ";
    }

    //toString method for field label, data, and output structure


    public static String toString(Job job) {
        String id = "\nID: " + job.getId();
        String name = "\nName: " + job.getName();
        String employer = "\nEmployer: " + job.getEmployer();
        String location = "\nLocation: " + job.getLocation();
        String positionType = "\nPosition Type: " + job.getPositionType();
        String coreCompetency = "\nCore Competency: " + job.getCoreCompetency();

//        String id = "\nID: " + id;
//        String name = "\nName: " + this.name;
//        String employer = "\nEmployer: " + this.employer;
//        String location = "\nLocation: " + this.location;
//        String positionType = "\nPosition Type: " + this.positionType;
//        String coreCompetency = "\nCore Competency: " + this.coreCompetency;

        if (name == "") {
            name = "\nName: \"Data Not Available\"";
        }
        if(employer == "") {
            employer = "\nEmployer: \"Data Not Available\"";
        }
        if(location == "") {
            location = "\nLocation: \"Data Not Available\"";
        }
        if (positionType == "") {
            positionType = "\nPosition Type: \"Data Not Available\"";
        }
        if (coreCompetency == "") {
            coreCompetency = "\nCore Competency: \"Data Not Available\"";
        }
        return (id + name + employer + location + positionType + coreCompetency + "\n");
    }

    //added equals() and hashCode() after adding getters/setters, equal set to id fields match

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (!(o instanceof Job)) return false;
        Job job = (Job) o;
        return getId() == job.getId();
    }

    @Override
    public int hashCode() {
        return Objects.hash(getId());
    }

    // Added getters for each field EXCEPT nextId. Add setters for each field EXCEPT nextID and id.


    public int getId() {
        return id;
    }

    public String getName() {
        return name;
    }

    public Employer getEmployer() {
        return employer;
    }

    public Location getLocation() {
        return location;
    }

    public PositionType getPositionType() {
        return positionType;
    }

    public CoreCompetency getCoreCompetency() {
        return coreCompetency;
    }

    public void setName(String name) {
        this.name = name;
    }

    public void setEmployer(Employer employer) {
        this.employer = employer;
    }

    public void setLocation(Location location) {
        this.location = location;
    }

    public void setPositionType(PositionType positionType) {
        this.positionType = positionType;
    }

    public void setCoreCompetency(CoreCompetency coreCompetency) {
        this.coreCompetency = coreCompetency;
    }

}
